+++ app/globals.css
@@ -1,3 +1,5 @@
+@import url('https://fonts.googleapis.com/css2?family=Lato:ital,wght@0,100;0,300;0,400;0,700;0,900;1,100;1,300;1,400;1,700;1,900&family=Over+the+Rainbow&display=swap');
+/* @import url('https://fonts.googleapis.com/css2?family=Lato:ital,wght@0,100;0,300;0,400;0,700;0,900;1,100;1,300;1,400;1,700;1,900&display=swap'); */
 @import url("https://fonts.googleapis.com/css2?family=Namdhinggo:wght@400;500;600;700;800&family=Pacifico&family=Satisfy&display=swap");
 @tailwind base;
 @tailwind components;
@@ -5,12 +7,11 @@
 
 @layer base {
   * {
-    font-family: "Namdhinggo", serif;
-    font-weight: 400;
-    font-style: normal;
+    
+    font-family: "Lato", sans-serif;
     @apply border-border;
   }
-  body {
+  body {;
     @apply bg-background text-foreground;
   }
   :root {
