+++ public/assets/HeroPic.tsx
@@ -1,0 +1,231 @@
+import React from 'react'
+
+type Props = {}
+
+const HeroPic = (props: Props) => {
+  return (
+    <svg
+        xmlns="http://www.w3.org/2000/svg"
+        viewBox="0 0 400 300"
+        data-imageid="music-flatline"
+        // imagename="Music"
+        className="illustrations_image sm:w-[450px] w-72 "
+        // style={{ width: 158 }}
+      >
+        <g id="_72_music" data-name="#72_music">
+          <path
+            d="M199.42,33.46c-.16-2.31-10.49,4.41-10.49,4.41s8.75-6.72,8-8.4-10.61,5.67-10.61,5.67,9-6.93,7.61-8.41-11,6.31-11,6.31,9.8-8.62,6.06-8.83c-1.68-.09-14.87,8.88-29.16,19.65,4,5,8,9.83,12.32,14.5,4.58-4.74,8.19-8.51,8.68-9.14,1.47-1.89,16.07-.63,16.13-2.94s-9.62-3.36-9.62-3.36S199.58,35.77,199.42,33.46Z"
+            fill="#fff"
+          />
+          <path
+            d="M172.09,58.81h0a.44.44,0,0,1-.33-.14c-3.84-4.16-7.76-8.77-12.35-14.53a.46.46,0,0,1-.1-.34.5.5,0,0,1,.18-.31c10.57-8,27.07-19.73,29.43-19.73h0a1.19,1.19,0,0,1,1.22.75c.32.9-.8,2.56-2.27,4.23,2.74-1.83,5.48-3.24,6.34-2.32a1,1,0,0,1,.26.81c-.07.82-1,2-2.24,3.31,1.88-1.11,3.64-1.95,4.45-1.74a.84.84,0,0,1,.6.47c.49,1.05-1.39,3.19-3.58,5.2,2.35-1.26,4.73-2.28,5.64-1.82a.91.91,0,0,1,.51.78h0c.15,2.16-7.8,7.16-11.21,9.2,3.12.43,8.73,1.5,8.68,3.66,0,1.57-3.31,1.83-8.66,2.11-3.12.17-7,.38-7.56,1.1s-3.3,3.58-8.71,9.18A.49.49,0,0,1,172.09,58.81ZM160.42,43.94c4.31,5.4,8,9.79,11.68,13.75,6.34-6.56,8-8.38,8.31-8.75.76-1,3.62-1.2,8.23-1.45,2.72-.15,7.78-.42,7.8-1.22,0-1.17-4.76-2.41-9.21-2.9a.45.45,0,0,1-.39-.36.46.46,0,0,1,.21-.49c5.93-3.44,12-7.76,11.92-9h0a.06.06,0,0,0,0,0c-.62-.43-5.1,1.76-9.79,4.81a.46.46,0,0,1-.52-.75c4.51-3.46,8.17-7.13,7.83-7.85-.67-.33-5.23,2.34-9.91,5.85a.46.46,0,0,1-.56-.73c2.87-2.22,7.46-6.25,7.58-7.62,0-.06,0-.1,0-.11-.56-.62-5.53,2.41-10.41,6.35a.47.47,0,0,1-.64-.07.44.44,0,0,1,0-.63c3.49-3.08,7.06-7,6.74-7.89,0-.09-.23-.13-.41-.14h0C187.51,24.67,176.3,32,160.42,43.94Z"
+            fill="#231f20"
+          />
+          <path
+            d="M121.29,78.18c-2.94,9.29,14.82,52.2,14.82,52.2l22-27.19c-.49-.84-11.61-20.18-11.61-20.18s16.41-15.19,25.56-24.65c-4.3-4.67-8.36-9.54-12.32-14.5C142.18,57.1,122.91,73.06,121.29,78.18Z"
+            fill="#231f20"
+          />
+          <path
+            d="M136.11,130.82H136a.43.43,0,0,1-.34-.27c-.73-1.75-17.79-43.15-14.83-52.49h0c1.67-5.27,21-21.3,38.63-34.54a.44.44,0,0,1,.61.07c4.57,5.74,8.48,10.34,12.31,14.48a.45.45,0,0,1,0,.6c-8.43,8.72-23.2,22.46-25.33,24.43,1.42,2.47,11,19.1,11.45,19.88a.46.46,0,0,1,0,.5l-22,27.18A.44.44,0,0,1,136.11,130.82Zm23.58-86.35c-16.6,12.52-36.38,28.8-38,33.84h0c-2.66,8.4,12.27,45.67,14.52,51.21l21.38-26.37L146.14,83.23a.46.46,0,0,1,.09-.55c.16-.15,16.16-15,25.26-24.33C167.81,54.35,164.05,49.93,159.69,44.47Zm-38.4,33.71h0Z"
+            fill="#231f20"
+          />
+          <path
+            d="M173.59,122.54c1.8,12.46,10.09,23,19.09,32.63,5.39,5.72,12,11.54,19.89,11.17,3.77-.19,7.31-1.81,10.6-3.65A80.32,80.32,0,0,0,244.94,145a4.88,4.88,0,0,1-4.87-4.36,9.64,9.64,0,0,0,10-1.12c6.12-4.86,3.15-9.15-1.53-12.77a56.43,56.43,0,0,1-16.95-21.41c-5.84-13.06-17.44-30.51-34.62-26.53-16.52,3.82-21,18.64-23.28,33.34A34.85,34.85,0,0,0,173.59,122.54Z"
+            fill="#fff"
+          />
+          <path
+            d="M211.83,166.81c-7.75,0-14.19-5.7-19.48-11.33-8.45-9-17.34-19.93-19.21-32.87a35.41,35.41,0,0,1,.07-10.58c1-6.3,2.59-14.37,6.36-20.91a25.68,25.68,0,0,1,17.26-12.81c7.84-1.81,22.82-.74,35.15,26.79a55.77,55.77,0,0,0,16.81,21.24c3.5,2.71,5.14,5.09,5.16,7.51,0,2-1.21,4.06-3.63,6a10.1,10.1,0,0,1-9.65,1.51,4.49,4.49,0,0,0,4.25,3.15.45.45,0,0,1,.43.26.44.44,0,0,1-.07.49,79.9,79.9,0,0,1-21.89,17.85c-3.09,1.72-6.8,3.51-10.8,3.7ZM201.7,78.67a20.83,20.83,0,0,0-4.66.53c-16.23,3.76-20.61,18.06-22.93,33a34.54,34.54,0,0,0-.07,10.3h0c1.83,12.7,10.62,23.5,19,32.38,5.32,5.66,11.79,11.4,19.53,11,3.81-.18,7.4-1.92,10.4-3.59a79.07,79.07,0,0,0,21-16.95,5.37,5.37,0,0,1-4.38-4.71.44.44,0,0,1,.2-.41.43.43,0,0,1,.45,0,9.18,9.18,0,0,0,9.49-1.06c2.19-1.75,3.3-3.52,3.28-5.27,0-2.08-1.58-4.3-4.8-6.79a56.75,56.75,0,0,1-17.08-21.58C221.24,83.36,209.7,78.67,201.7,78.67Z"
+            fill="#231f20"
+          />
+          <path
+            d="M287.82,161.65c-.7,2.77-12.19,23.29-12.19,23.29l15,8.11s13.57-22.74,14.73-24.25,11-14.62,12.34-19.22-8.12-8-12-7.63S288.52,158.88,287.82,161.65Z"
+            fill="#fff"
+          />
+          <path
+            d="M290.61,193.51a.39.39,0,0,1-.21-.06l-15-8.11a.44.44,0,0,1-.22-.28.43.43,0,0,1,0-.35c4-7.07,11.59-21,12.14-23.17h0c.74-2.94,14.19-19.69,18.25-20,2.95-.25,9.1,1.54,11.57,4.53a3.93,3.93,0,0,1,.92,3.68c-1.43,5-12.31,19.24-12.41,19.38-1.14,1.47-14.57,24-14.71,24.2A.43.43,0,0,1,290.61,193.51Zm-14.35-8.76,14.19,7.69c1.83-3.07,13.44-22.5,14.53-23.92,2.29-3,11.06-14.88,12.26-19.07a3,3,0,0,0-.75-2.85c-2.25-2.73-8.1-4.43-10.78-4.2-3.66.32-16.84,17-17.45,19.36C287.6,164.37,277.92,181.78,276.26,184.75Z"
+            fill="#231f20"
+          />
+          <path
+            d="M283.51,170.52l-13.4,16.34-28.9-19.37L228.9,190.41s40.93,30.73,46.73,27.79S299,179,299,179Z"
+            fill="#231f20"
+          />
+          <path
+            d="M274.7,218.82c-8.44,0-42.07-25.06-46.07-28.06a.45.45,0,0,1-.12-.56l12.31-22.92a.44.44,0,0,1,.28-.22.43.43,0,0,1,.35.06L270,186.27l13.15-16a.43.43,0,0,1,.55-.11l15.54,8.48a.45.45,0,0,1,.18.58c-.72,1.49-17.73,36.43-23.61,39.4A2.4,2.4,0,0,1,274.7,218.82Zm-45.24-28.54c11.56,8.65,41.45,29.82,46,27.53,5.29-2.69,21.13-34.76,23-38.62l-14.84-8.11-13.17,16.06a.42.42,0,0,1-.58.09l-28.51-19.11Z"
+            fill="#231f20"
+          />
+          <polygon
+            points="295.78 130.56 295.65 170.52 314.56 170.52 314.56 130.56 295.78 130.56"
+            fill="#231f20"
+          />
+          <path
+            d="M314.56,171H295.65a.44.44,0,0,1-.44-.45l.12-40a.45.45,0,0,1,.45-.44h18.78a.44.44,0,0,1,.44.44v40A.44.44,0,0,1,314.56,171Zm-18.46-.88h18V131h-17.9Z"
+            fill="#231f20"
+          />
+          <polygon
+            points="297.7 134.09 297.6 165.49 312.62 165.49 312.62 134.09 297.7 134.09"
+            fill="#fff"
+          />
+          <path
+            d="M312.62,165.94h-15a.44.44,0,0,1-.32-.13.44.44,0,0,1-.13-.33l.09-31.39a.46.46,0,0,1,.46-.45h14.92a.45.45,0,0,1,.45.45v31.4A.45.45,0,0,1,312.62,165.94ZM298.06,165h14.1V134.55h-14Z"
+            fill="#231f20"
+          />
+          <path
+            d="M287,163.73s-2.82-13.46,1.41-15.57,18.77-4.58,16.52-1.28-10.13,5.34-10.13,5.34,1.62,4.71-.16,8.05"
+            fill="#fff"
+          />
+          <path
+            d="M287,164.18a.44.44,0,0,1-.44-.36c-.12-.56-2.85-13.83,1.65-16.07,3.65-1.81,15.43-4.14,17.15-2.18a1.19,1.19,0,0,1-.06,1.57c-2,3-8.17,4.9-9.93,5.4.37,1.31,1.21,5.06-.33,8a.46.46,0,0,1-.62.18.47.47,0,0,1-.19-.62c1.67-3.12.15-7.63.13-7.68a.45.45,0,0,1,0-.36.48.48,0,0,1,.29-.23c.08,0,7.74-2,9.87-5.15.15-.22.19-.38.12-.46-1.08-1.24-12,.4-16.06,2.4-3.21,1.6-1.92,11.41-1.16,15.06a.45.45,0,0,1-.35.54Z"
+            fill="#231f20"
+          />
+          <path
+            d="M182.45,137.59,157.54,94.83l-28,31.09,10.4,22.42s-16,37.87-16,112.1c0,0,85.18-.37,106.62-.37l-28-43.55s7-20.18,10.61-25.22l20.85,11.35,24.38-33.21Z"
+            fill="#000000"
+            className="target-color"
+          />
+          <path
+            d="M182.09,178.35c13.75,8,17-24.52,17.61-31.56a14.5,14.5,0,0,0-.34-5.8c-.79-2.4-2.64-4.26-4.49-6-2.21-2-4.76-4.1-7.77-4.23a9,9,0,0,0-7.17,3.75,22,22,0,0,0-3.44,7.62c-3.24,11.31-3.12,23.25,3.13,33.53a7.19,7.19,0,0,0,2.17,2.47Z"
+            fill="#231f20"
+          />
+          <path
+            d="M186.2,180.05a8.64,8.64,0,0,1-4.33-1.32l-.31-.19a7.53,7.53,0,0,1-2.31-2.62c-5.73-9.43-6.83-21.14-3.18-33.88a22.45,22.45,0,0,1,3.51-7.77,9.37,9.37,0,0,1,7.54-3.92c3.08.13,5.69,2.17,8,4.35,1.88,1.73,3.78,3.64,4.61,6.15a15.27,15.27,0,0,1,.36,6c-.93,10.49-3.6,28.66-11,32.49A6.23,6.23,0,0,1,186.2,180.05Zm.62-48.82a8.58,8.58,0,0,0-6.55,3.58,21.62,21.62,0,0,0-3.35,7.47c-3.59,12.5-2.52,24,3.08,33.18a6.73,6.73,0,0,0,2,2.33l.29.17h0c2.38,1.39,4.48,1.58,6.41.58,5.14-2.66,9-14.24,10.54-31.79a14.35,14.35,0,0,0-.32-5.63c-.76-2.31-2.57-4.12-4.37-5.78-2.24-2.06-4.69-4-7.49-4.11Z"
+            fill="#231f20"
+          />
+          <path
+            d="M223.12,117.37c1.27,4.61-.81,10.48-3.93,13.89-2.08,2.25-6.64,2.85-7.56-.81A18.57,18.57,0,0,1,214,116.22c2-3.29,5.83-5.17,8.29-.94A8.71,8.71,0,0,1,223.12,117.37Z"
+            fill="#231f20"
+          />
+          <path
+            d="M215.05,133.51a4.8,4.8,0,0,1-1.14-.13,3.7,3.7,0,0,1-2.71-2.83A19.09,19.09,0,0,1,213.58,116c1.4-2.25,3.38-3.61,5.23-3.6,1.05,0,2.56.49,3.82,2.67a9.47,9.47,0,0,1,.92,2.19h0c1.38,5-1,11-4,14.3A6.24,6.24,0,0,1,215.05,133.51Zm3.71-20.23c-1.52,0-3.21,1.21-4.43,3.17a18.17,18.17,0,0,0-2.27,13.89,2.79,2.79,0,0,0,2.07,2.18,5.19,5.19,0,0,0,4.73-1.56c2.82-3.06,5.12-8.83,3.84-13.47h0a8.42,8.42,0,0,0-.83-2c-.85-1.46-1.88-2.21-3.07-2.22Z"
+            fill="#231f20"
+          />
+          <path
+            d="M201.67,138c-2,7.67-4,25.14-16.75,16a16.18,16.18,0,0,1-3.38-3.62c-6.28-8.37-4.49-13.65-1.24-22.84,1.4-4,2.61-8,4-12s3.68-7.76,5.58-11.55c-.07.13,6.25,1,12.57-1.87S214,96.27,214,96.27c1.71,2.56,3.56,4.5,4.47,7.54a23,23,0,0,1,.71,9.08c-.8,7.47-3.79,15.55-7.18,22.22S201.67,138,201.67,138Z"
+            fill="#fff"
+          />
+          <path
+            d="M191.36,157.07c-2,0-4.23-.9-6.71-2.68a16.71,16.71,0,0,1-3.48-3.72c-6.47-8.63-4.52-14.14-1.3-23.26.67-1.92,1.31-3.89,1.93-5.79.67-2.05,1.36-4.16,2.1-6.23a66,66,0,0,1,3.95-8.43c.56-1.06,1.12-2.12,1.65-3.18h0a.45.45,0,0,1,.61-.2h0a23,23,0,0,0,12.15-1.88c6.22-2.85,11.46-5.81,11.51-5.84a.47.47,0,0,1,.61.15c.43.64.86,1.24,1.29,1.83a18.82,18.82,0,0,1,3.24,5.82,23.71,23.71,0,0,1,.73,9.27c-.73,6.69-3.29,14.64-7.23,22.37a7.3,7.3,0,0,1-4.33,3.93,9.27,9.27,0,0,1-6.13-.55c-.2.79-.41,1.68-.62,2.61-1.32,5.73-3.12,13.59-7.89,15.39A6,6,0,0,1,191.36,157.07Zm-1.2-52.57c-.48,1-1,1.92-1.5,2.88a65.52,65.52,0,0,0-3.9,8.32c-.73,2-1.42,4.16-2.09,6.2s-1.26,3.88-1.94,5.81c-3.23,9.14-5,14.17,1.18,22.41a15.87,15.87,0,0,0,3.28,3.53c3.14,2.26,5.82,3,8,2.18,4.31-1.63,6-9.2,7.32-14.73.27-1.16.51-2.25.76-3.18a.47.47,0,0,1,.66-.29,8.72,8.72,0,0,0,6,.73,6.37,6.37,0,0,0,3.77-3.46c3.89-7.64,6.43-15.48,7.14-22.06a22.57,22.57,0,0,0-.7-8.9A18.09,18.09,0,0,0,215,98.39c-.36-.49-.73-1-1.08-1.51-1.31.72-5.91,3.23-11.22,5.66A24,24,0,0,1,190.16,104.5Z"
+            fill="#231f20"
+          />
+          <path
+            d="M199.57,145.42s-6.19-1.41-9-6.56-1.11-9.1-1.11-9.1,5.4,6.25,12.23,8.27Z"
+            fill="#231f20"
+          />
+          <path
+            d="M199.57,145.86h-.1c-.26-.06-6.42-1.51-9.3-6.77s-1.21-9.32-1.14-9.49a.44.44,0,0,1,.34-.26.46.46,0,0,1,.4.14c0,.06,5.41,6.18,12,8.14a.44.44,0,0,1,.3.55L200,145.54A.44.44,0,0,1,199.57,145.86Zm-9.94-15.23a11.16,11.16,0,0,0,1.31,8c2.29,4.18,6.93,5.82,8.32,6.23l1.87-6.55C195.6,136.54,191.12,132.2,189.63,130.63Z"
+            fill="#231f20"
+          />
+          <path
+            d="M203.36,125.4a8.32,8.32,0,0,1-3.57-1.12.44.44,0,0,1,.44-.77c.77.44,3,1.44,4,.79.65-.42-.06-4.53-1.56-9a.44.44,0,0,1,.84-.28c.9,2.67,2.81,9,1.19,10A2.5,2.5,0,0,1,203.36,125.4Z"
+            fill="#231f20"
+          />
+          <path
+            d="M201.53,130.82a6.68,6.68,0,0,1-3-.66c-2.5-1.24-4-4.15-4.52-8.65a.45.45,0,0,1,.39-.49.44.44,0,0,1,.49.39c.47,4.18,1.82,6.86,4,8,3.33,1.65,7.58-.85,7.62-.87a.44.44,0,0,1,.61.15.45.45,0,0,1-.15.61A12,12,0,0,1,201.53,130.82Z"
+            fill="#231f20"
+          />
+          <path
+            d="M202,110.86a.44.44,0,0,1-.44-.41c-.07-1-.36-1.61-.88-1.87-1-.52-2.87.42-3.49.83a.44.44,0,0,1-.49-.74c.11-.07,2.67-1.73,4.37-.88a2.85,2.85,0,0,1,1.37,2.6.44.44,0,0,1-.41.47Z"
+            fill="#231f20"
+          />
+          <path
+            d="M213.09,115.82a.44.44,0,0,1-.44-.4s-.2-2.2-1.45-3c-1.08-.71-2.34.91-2.35.93a.44.44,0,0,1-.62.08.44.44,0,0,1-.09-.62c.61-.8,2.13-2.06,3.55-1.13,1.61,1.06,1.83,3.58,1.84,3.68a.44.44,0,0,1-.4.48Z"
+            fill="#231f20"
+          />
+          <path
+            d="M177.92,113.4a8.89,8.89,0,0,0,.8,4.65c1.42,2.63,4.92,3.7,7.78,2.85,7.47-2.21,11.06-16.42,5.41-22-5.25-5.18-11.21,3-12.44,7.83C178.9,109,178,111.1,177.92,113.4Z"
+            fill="#231f20"
+          />
+          <path
+            d="M184.5,121.63a7,7,0,0,1-6.17-3.37,9.11,9.11,0,0,1-.85-4.88h0a21.22,21.22,0,0,1,1-4.92c.18-.61.36-1.22.52-1.83.87-3.36,4.1-8.57,7.91-9.62a5.26,5.26,0,0,1,5.27,1.58c2.73,2.69,3.62,7.56,2.33,12.71s-4.32,9-7.92,10A7.77,7.77,0,0,1,184.5,121.63Zm-6.14-8.21a8.36,8.36,0,0,0,.75,4.43,6.35,6.35,0,0,0,7.27,2.63c3.31-1,6.11-4.58,7.31-9.39s.41-9.4-2.09-11.87a4.39,4.39,0,0,0-4.42-1.37c-3.44,1-6.49,5.93-7.29,9-.16.62-.34,1.24-.52,1.86a20.35,20.35,0,0,0-1,4.71Z"
+            fill="#231f20"
+          />
+          <path
+            d="M219.45,122.19a1,1,0,0,1-.53-.15,1,1,0,0,1-.33-1.37c.11-.18,11.2-18.42,13.1-29.79.69-4.12-1.14-8.3-5.15-11.79-5.48-4.76-13.66-6.89-19.45-5.06-10.3,3.26-15.78,27-15.84,27.2a1,1,0,0,1-1.19.75,1,1,0,0,1-.76-1.19c.23-1,5.8-25.07,17.19-28.67,6.51-2.05,15.29.19,21.36,5.46,4.56,4,6.63,8.8,5.82,13.63-2,11.76-12.91,29.74-13.37,30.5A1,1,0,0,1,219.45,122.19Z"
+            fill="#231f20"
+          />
+          <path
+            d="M273.67,263.82a34.09,34.09,0,0,1-10.55-1.71,34.61,34.61,0,0,1-11.07-6.44c-5.55-4.49-10.44-9.85-15.17-15-3-3.29-6.1-6.69-9.34-9.83-10.15-9.84-20.23-15.08-30-15.56-13.14-.66-24.79,6.85-35.41,15.06-1,.8-2.08,1.63-3.14,2.47-8.59,6.81-18.33,14.52-30.33,12.71-8.18-1.23-14.53-6.87-18.41-11.38-9.55-11.11-14-27-11.79-42.41,2.34-16.6,11.59-28.3,27.5-34.78,13.17-5.36,25.6-12.86,37.63-20.11l.1-.05c7.45-4.5,15.36-10.26,17.49-18.9a.5.5,0,1,1,1,.24c-2.22,9-10.32,14.91-17.95,19.51l-.09.06c-12.06,7.27-24.53,14.79-37.77,20.18-15.56,6.33-24.6,17.76-26.89,34C97.34,207,101.66,222.55,111,233.44c3.77,4.39,9.92,9.87,17.8,11.05,11.57,1.75,20.72-5.5,29.56-12.5l3.15-2.48c10.78-8.33,22.61-15.95,36.07-15.27,10,.5,20.3,5.83,30.63,15.84,3.26,3.17,6.37,6.58,9.38,9.88,4.7,5.16,9.56,10.49,15.06,14.93a33.73,33.73,0,0,0,10.74,6.27c13.51,4.36,27,.07,40.19-12.76,19.57-19.08,30.6-48,30.25-79.44-.16-14.68-2.9-26.15-8.38-35.08-2.21-3.61-5-7.63-9.54-9.64-2.32-1-7.76-1.38-9.86,1.57-.8,1.13-.88,2.46-1,4.15,0,.73-.1,1.55-.21,2.44a.5.5,0,0,1-1-.13c.12-.86.16-1.63.21-2.37.11-1.78.21-3.31,1.18-4.67,2.44-3.43,8.42-3.07,11.08-1.9,4.84,2.12,7.69,6.29,10,10,5.58,9.09,8.36,20.73,8.53,35.6.35,31.68-10.78,60.9-30.55,80.16C294.27,258.9,284,263.81,273.67,263.82Z"
+            fill="#231f20"
+          />
+          <path
+            d="M177,43.42a.52.52,0,0,1-.38-.17.51.51,0,0,1,0-.71c2.2-1.91,4.38-2.75,6.49-2.51a7,7,0,0,1,4.54,2.59.51.51,0,0,1-.11.7.5.5,0,0,1-.7-.11A6.08,6.08,0,0,0,183,41c-1.82-.21-3.74.56-5.72,2.28A.49.49,0,0,1,177,43.42Z"
+            fill="#231f20"
+          />
+          <path
+            d="M206.79,205.32a.52.52,0,0,1-.21,0,.47.47,0,0,1-.2-.62L219,179.4a.45.45,0,0,1,.81.41L207.2,205.07A.45.45,0,0,1,206.79,205.32Z"
+            fill="#231f20"
+          />
+          <path
+            d="M183.28,210.25c-14.45,0-31.43-4.68-32.37-4.94a.46.46,0,1,1,.24-.88c.31.08,30.63,8.45,44.65,3.1a.46.46,0,0,1,.59.27.44.44,0,0,1-.26.58A37.1,37.1,0,0,1,183.28,210.25Z"
+            fill="#231f20"
+          />
+          <rect x="303.68" y="127.76" width="2.18" height="3.24" fill="#231f20" />
+          <path
+            d="M305.86,131.44h-2.18a.45.45,0,0,1-.45-.44v-3.24a.44.44,0,0,1,.45-.44h2.18a.44.44,0,0,1,.44.44V131A.44.44,0,0,1,305.86,131.44Zm-1.74-.89h1.3V128.2h-1.3Z"
+            fill="#231f20"
+          />
+          <path
+            d="M186.65,118.32l-.72-.7a14.57,14.57,0,0,0,2.57-17.2l.86-.51A15.6,15.6,0,0,1,186.65,118.32Z"
+            fill="#fff"
+          />
+          <path
+            d="M219.65,131.44l-.93-.37c2.92-7.53,2.18-13.09,2.18-13.14l1-.14C221.92,118,222.67,123.64,219.65,131.44Z"
+            fill="#fff"
+          />
+          <rect
+            x="137.72"
+            y="86.21"
+            width="10.36"
+            height={1}
+            transform="translate(-19.01 128.24) rotate(-45.63)"
+            fill="#fff"
+          />
+          <rect
+            x="273.46"
+            y="184.46"
+            width={1}
+            height="8.57"
+            transform="translate(-16.03 351.8) rotate(-63.91)"
+            fill="#fff"
+          />
+          <path
+            d="M69.78,137.33c-.42-.1-.87-.17-1.28-.31a3.82,3.82,0,0,1-2.74-2.61,4,4,0,0,1,.6-3.25A7.27,7.27,0,0,1,71.18,128c1.29-.29,1.87-.22,3.92.28V104.2l23.68-4.74V100c0,9.64,0,19.27,0,28.91a5,5,0,0,1-1.75,4,7.38,7.38,0,0,1-7.23,1.78,3.67,3.67,0,0,1-2.74-2.6,3.88,3.88,0,0,1,.59-3.24,7.21,7.21,0,0,1,4.84-3.16,6.86,6.86,0,0,1,3.88.28V109.41l-1.68.33q-8.45,1.7-16.87,3.37c-.31.06-.36.18-.36.46,0,5.86,0,11.71,0,17.56a5.37,5.37,0,0,1-2.31,4.59,7.92,7.92,0,0,1-3.78,1.54,1.72,1.72,0,0,0-.21.07Z"
+            fill="#e6e7e8"
+          />
+          <path
+            d="M262.33,65.29a5,5,0,0,0,.6-.32,1.93,1.93,0,0,0,1-1.65,2,2,0,0,0-.72-1.53,3.71,3.71,0,0,0-2.79-.92,3.51,3.51,0,0,0-1.9.64l-3.15-11.92L243,50.33l.07.25c1.25,4.77,2.52,9.54,3.77,14.32a2.56,2.56,0,0,0,1.38,1.75,3.79,3.79,0,0,0,3.81-.07,1.87,1.87,0,0,0,1-1.64,2,2,0,0,0-.71-1.53,3.74,3.74,0,0,0-2.82-.93,3.51,3.51,0,0,0-1.88.64L245.52,55l.88-.06,8.79-.53c.16,0,.2,0,.24.18.76,2.89,1.54,5.79,2.28,8.69a2.75,2.75,0,0,0,1.74,2,4.1,4.1,0,0,0,2.08.27h.11Z"
+            fill="#e6e7e8"
+          />
+          <path
+            d="M310.44,228.17c0-.19,0-.38,0-.57a3.06,3.06,0,0,0-.57-1.45,6.36,6.36,0,0,0-1.93-1.83,7.11,7.11,0,0,0-2.51-1,4.52,4.52,0,0,0-1.72,0l-.15,0a1.21,1.21,0,0,1,0-.2c.1-.39.2-.77.31-1.14l2-7c.73-2.54,1.47-5.08,2.2-7.62l0-.15-.7-.06a12.44,12.44,0,0,0-5.59,1,8.07,8.07,0,0,0-3.91,3.43c-.25.44-.44.91-.65,1.38-.14.3-.25.62-.38.93a.51.51,0,0,0,0,.13l.25-.15a19,19,0,0,1,2.49-1.38,8.88,8.88,0,0,1,4.25-.76,7.77,7.77,0,0,1,1.84.41l0,0a.83.83,0,0,0,0,.15q-1.82,6.32-3.61,12.63a.58.58,0,0,0,0,.23,3.61,3.61,0,0,0,.9,2,6.59,6.59,0,0,0,2.34,1.81l.78.34.72.21H307a4.55,4.55,0,0,0,1.8,0,2.1,2.1,0,0,0,1.29-.7,5.4,5.4,0,0,0,.3-.47Z"
+            fill="#e6e7e8"
+          />
+          <path
+            d="M65.79,230.37c0-.24,0-.49.11-.73a3.92,3.92,0,0,1,1-1.75,8.37,8.37,0,0,1,2.86-1.95,9.16,9.16,0,0,1,3.42-.77,6,6,0,0,1,2.21.35l.19.07a2.28,2.28,0,0,0,0-.26c0-.51-.1-1-.16-1.52q-.57-4.73-1.15-9.43c-.42-3.42-.85-6.83-1.27-10.24l0-.2c.31,0,.62,0,.91.07a16.24,16.24,0,0,1,7,2.39,10.52,10.52,0,0,1,4.31,5.2,17.93,17.93,0,0,1,.55,1.9c.12.42.2.85.3,1.28a.32.32,0,0,1,0,.17l-.29-.25a24.87,24.87,0,0,0-2.92-2.28,11.37,11.37,0,0,0-5.29-1.83,10.06,10.06,0,0,0-2.45.14l0,0a1.23,1.23,0,0,1,0,.19q1,8.48,2.05,17a.71.71,0,0,1,0,.29,4.71,4.71,0,0,1-1.57,2.43,8.4,8.4,0,0,1-3.38,1.84c-.35.11-.71.19-1.06.28l-1,.12H70a5.88,5.88,0,0,1-2.32-.32,2.66,2.66,0,0,1-1.5-1.15,5.06,5.06,0,0,1-.3-.67Z"
+            fill="#e6e7e8"
+          />
+          <path
+            d="M271.45,107.85c.13-.26.24-.53.39-.77a4.79,4.79,0,0,1,1.84-1.59,10.1,10.1,0,0,1,4-1.08,11,11,0,0,1,4.16.47,6.93,6.93,0,0,1,2.36,1.27l.19.15a1.78,1.78,0,0,0,.12-.29c.15-.59.29-1.18.42-1.78l2.4-11.1q1.29-6,2.57-12.06l.05-.24c.35.15.68.28,1,.43a19.42,19.42,0,0,1,6.94,5.44,12.54,12.54,0,0,1,2.84,7.57,23.12,23.12,0,0,1-.12,2.35c0,.53-.11,1-.17,1.57a.43.43,0,0,1-.06.18l-.24-.39a29,29,0,0,0-2.4-3.71,13.52,13.52,0,0,0-5.26-4.15,11.48,11.48,0,0,0-2.82-.79s0,0-.07,0a1.55,1.55,0,0,1-.05.23l-4.32,19.94a1,1,0,0,1-.13.33,5.66,5.66,0,0,1-2.72,2.13,10.23,10.23,0,0,1-4.54.75c-.44,0-.87-.06-1.31-.1l-1.15-.24-.19-.08a7,7,0,0,1-2.49-1.27,3.31,3.31,0,0,1-1.25-1.89,8.28,8.28,0,0,1-.07-.87Z"
+            fill="#e6e7e8"
+          />
+          <path
+            d="M95.17,153.39c-5.15,0-5.16,8,0,8S100.33,153.39,95.17,153.39Z"
+            fill="#e6e7e8"
+          />
+          <path
+            d="M277.86,137.15a2.11,2.11,0,0,0-.51-.88l-.41-.31a2,2,0,0,0-1-.28h-.56a1.46,1.46,0,0,0-.77.18,1.44,1.44,0,0,0-.65.41,1.56,1.56,0,0,0-.41.64,1.47,1.47,0,0,0-.17.77l.07.54a2.06,2.06,0,0,0,.51.88l.41.31a2,2,0,0,0,1,.27h.56a1.45,1.45,0,0,0,.77-.17,1.6,1.6,0,0,0,.65-.41,2.26,2.26,0,0,0,.58-1.42C277.91,137.51,277.88,137.33,277.86,137.15Z"
+            fill="#e6e7e8"
+          />
+          <path
+            d="M110.94,55.78c-3.22,0-3.23,5,0,5S114.16,55.78,110.94,55.78Z"
+            fill="#e6e7e8"
+          />
+          <path
+            d="M257.35,232c-3.22,0-3.22,5,0,5S260.57,232,257.35,232Z"
+            fill="#e6e7e8"
+          />
+        </g>
+      </svg>
+  )
+}
+
+export default HeroPic;
